renamed `drivers` module to `strategies` and the `Driver` base class to `Strategy`
