--- mesa/drivers/__init__.py
@@ -1,3 +1,0 @@
-from mesa.drivers.base import Driver, Optimizer
-from mesa.drivers.gpo import GPOptimizer
-from mesa.drivers.genetic import GeneticOptimizer
