+++ mesa/strategies/__init__.py
@@ -1,0 +1,3 @@
+from mesa.strategies.base import Strategy, Optimizer
+from mesa.strategies.gpo import GPOptimizer
+from mesa.strategies.genetic import GeneticOptimizer
