+++ tests/test_vsim.py
@@ -37,7 +37,7 @@ simulation = VSim(
 #   driver settings
 # ----------------------------------------------------------------------------
 
-from mesa.drivers.genetic import GeneticOptimizer
+from mesa.strategies.genetic import GeneticOptimizer
 
 # parameters to fix or vary. To fix set as a single value. To vary set bounds as a tuple
 params = {
